Finish common set of keys
